requirements.txt
python-binance==1.0.19
python-dotenv==1.0.1
websockets==12.0
rich==13.7.1
